Base optimization principles, through algebra delegation.
This commit poses the bases of the optimization. Unlike Alf,
optimization will apply on the fly, as expressions are built.

The commit introduces the Algebra module which has one method
per algebra operator. The default implementation instantiates
the operator classes, which is a non-optimization in practice,
but a valid implementation of the algebra. Specific Algebra
instantiations MAY override operator methods to provide real
optimizations in practice.

For this, Relation becomes a module, no longer a class, and
includes Algebra itself. Relation::Leaf is introduced to handle
tuple enumerables that do not expose the Relation interface.
And Operator includes Relation. Hence the distinction between
a logical-level Relation and physical-level Operator implem
is blured a little bit. This comes from Alf's experience,
where maintaining both concepts separate made type inference
and optimization particularly difficult.

When it comes to optimization, each Relation instantiation
may define its own optimization patterns, following a pure
object-oriented delegation pattern. An example is implemented
that optimizes restrictions of restrictions as predicate
conjunctions.