Refactor the optimization protocol towards DRYness.
The protocol is slighlty improved to avoid having to repeat type
inference and coercion mechanisms in all subclasses of Algebra.

The latter now implements type inference and coercions mechanisms
and delegates to an underscore version of the algebra methods that
takes the resulting type as parameter and requires strongly typed
arguments (e.g. predicate).