Optimize constants.restrict, push down restrictions.
This commit contributes an optimization when constants are
restricted. The usual case is to push the restriction down the
tree, when it does not refer to the attributes are known to be
constants.

Still, interresting cases arise when the predicate actually
refer to the constant attributes. Edge cases even allow to
return an empty relation, if it is statically known that the
predicate will evaluate to false.