Start support for Relvar updating: insert, update, delete.
This commit contributes basic algorithms for delegating insert,
update and delete invocations to the underlying database handler,
according to the Operator at hand.

The three methods are implemented on operators yielding not too
much update ambiguities:

* allbut
* constants
* extend
* project
* rename

An InvalidUpdateError may be raised when something wrong is
detected, such as violating the Relvar type and/or predicate.

Since we only have Sequel::Relation for now, the delegation
mechanism, when succeeding, ends up calling Sequel's own delete,
insert and update methods.