Revisit Sequel module to rely on valid SQL translation.
This commit enhances the whole library to ensure that SQL
translations made by the Sequel contribution are correct
with respect to the algebra semantics.

The original problem is that Sequel's Dataset instances
implement a SQL tree, not relation algebra. Therefore,
implementing the algebra and optimization rules is
semantically incorrect if special care is not taken.

The commit therefore introduces a Sql module, taken from
Alf, that implements a Bmg::Sql::Relation class supporting
correct SQL translation to an abstract syntax tree.

Bmg::Sequel::Relation extends Bmg::Sql::Relation and
translates that abstract syntax tree to Sequel::Dataset.

Integration tests are introduced. The whole mechanism
needs to be completed later.