Make sure to always unspy all operands, on all operators.
This commit fixes a bug where right operands of matching were
not unspied, leading to unoptimal query optimization in practice.

To avoid the situation from occuring again, the Spied relation
if refactored to avoid relying on a manually-maintained array
of operator names and a manually-maintained list of binary
methods.

We now use public methods of Algebra as the true and (supposed)
list of algebraic methods available.