Add NotMatching operator, with restrict optimization & SQL compilation.
This commit adds the inverse of Matching, an operator that filters
tuples of left operand to those for which no matching tuple exists
in right operand.

Restrict can always be pushed on the left operand, and never on right,
an optimization that is implemented. SQL compilation is implemented
too and integration tested.