Add a cache for indexes information loaded from sql database.
Sequel has a cache mechanism, that only caches the table schema,
not the indexes queries we do here. This commit adds a cache level
in Bmg itself, through a monkey patch extension of Sequel::Database
to profit from thread safety given by Sequel.synchronize.

The cache is only enabled when the standard cache_schema option
in Sequel is true.