Improve SQL generation of JOINs through linearization of clauses.
This commit adds a FromClauseOrderer helper, that transforms a
`from_clause` AST to a list of join clauses. The tool is used
by the SQL compiler and Sequel translator to avoid unnecessary
subqueries and WITH clauses.

It is still an open question whether the algorithm is fully
correct, since a NotImplementedError can sometimes be raised.