Add an option to the image operator to preserve joining attributes in the result.
Sometimes, it appears to be convenient to preserve the `on` attributes in the
resulting attribute of the image operator. For instance, one may need to do
`left.image(right, as, on)` and then may expect the new `as` attribute to
contain the `on` attributes of the `right` relation, in order to be able to
filter the `on` attributes from the resulting relation with a project or an
allbut operation.

This commits adds to the image operator a `preserve` option which applies this
behavior when true.
Default behavior unchanged.