Fix SQL generation when joining with a summarize.
The current compilateur raises an error or generates wrong SQL.
This first commit (incomplete) prevents such joins and forces
a WITH expression for the summary part of the join.

This commit is not correct/complete per se since some summaries
yield not GROUP BY clause, so the check to detected unjoinability
is actually wrong.